Refactor : MyParty 디렉토리 생성
MyParty도 디렉토리를 이용해 관련 컴포넌트를 분리했습니다.
