+++ src/components/MyParty/MyClosedparties.jsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
 import React from 'react';
 import PropTypes from 'prop-types';
-import MyParty from './MyParty';
-import { SIZES, PARTY_COLORS } from '../styles/constants';
+import MyClosedParty from './MyClosedParty';
+import { SIZES } from '../../styles/constants';
 
 const Container = styled.div``;
 
@@ -33,32 +33,20 @@ const PartyContainer = styled.div`
   }
 `;
 
-function MyParties({ parties }) {
+function MyClosedParties({ parties }) {
   return (
     <Container>
       <PartyContainer>
         {parties.map((party) => {
-          const color = PARTY_COLORS[Math.floor(Math.random() * PARTY_COLORS.length)];
-          return (
-            <MyParty
-              key={party.id}
-              title={party.title}
-              hashtags={party.hashtags}
-              status={party.status}
-              joinNumber={party.joinNumber}
-              goalNumber={party.goalNumber}
-              id={party.id}
-              color={color}
-            />
-          );
+          return <MyClosedParty key={party.id} title={party.title} />;
         })}
       </PartyContainer>
     </Container>
   );
 }
 
-MyParties.propTypes = {
+MyClosedParties.propTypes = {
   parties: PropTypes.arrayOf(PropTypes.object),
 };
 
-export default MyParties;
+export default MyClosedParties;
